refactor: replace uint with uint256 in the smart contracts

+++ kickstart/ethereum/contracts/Campaign.sol
@@ -3,12 +3,12 @@ pragma solidity >=0.5.0 <0.7.0;
 contract CampaignFactory {
     Campaign[] public deployedCampaigns;
 
-    function createCampaign(uint minimum) public {
+    function createCampaign(uint256 minimum) public {
         Campaign newCampaign = new Campaign(minimum, msg.sender);
         deployedCampaigns.push(newCampaign);
     }
 
-    function getDeployedCampaigns() public view returns(Campaign[] memory) {
+    function getDeployedCampaigns() public view returns (Campaign[] memory) {
         return deployedCampaigns;
     }
 }
@@ -16,20 +16,20 @@ contract CampaignFactory {
 contract Campaign {
     struct Request {
         string description;
-        uint value;
+        uint256 value;
         address payable recipient;
         bool complete;
-        uint approvalCount;
+        uint256 approvalCount;
         mapping(address => bool) approvals;
     }
 
     address public manager;
-    uint public minimumContribution;
+    uint256 public minimumContribution;
     mapping(address => bool) public approvers;
-    uint public approversCount;
+    uint256 public approversCount;
     Request[] public requests;
 
-    constructor(uint minimum, address creator) public {
+    constructor(uint256 minimum, address creator) public {
         manager = creator;
         minimumContribution = minimum;
     }
@@ -43,18 +43,23 @@ contract Campaign {
         approversCount++;
     }
 
-    function createRequest(string memory description, uint value, address payable recipient) public onlyManager {
-        Request memory newRequest = Request({
-            description: description,
-            value: value,
-            recipient: recipient,
-            complete: false,
-            approvalCount: 0
-        });
+    function createRequest(
+        string memory description,
+        uint256 value,
+        address payable recipient
+    ) public onlyManager {
+        Request memory newRequest =
+            Request({
+                description: description,
+                value: value,
+                recipient: recipient,
+                complete: false,
+                approvalCount: 0
+            });
         requests.push(newRequest);
     }
 
-    function approveRequest(uint index) public {
+    function approveRequest(uint256 index) public {
         Request storage request = requests[index];
         require(
             approvers[msg.sender],
@@ -69,22 +74,29 @@ contract Campaign {
         request.approvalCount++;
     }
 
-    function finalizeRequest(uint index) public onlyManager {
+    function finalizeRequest(uint256 index) public onlyManager {
         Request storage request = requests[index];
         require(
             request.approvalCount > (approversCount / 2),
             "This request needs more approvals before it can be finalized"
         );
-        require(
-            !(request.complete),
-            "This request has already been finalized"
-        );
+        require(!(request.complete), "This request has already been finalized");
 
         request.recipient.transfer(request.value);
         request.complete = true;
     }
 
-    function getSummary() public view returns (uint, uint, uint, uint, address) {
+    function getSummary()
+        public
+        view
+        returns (
+            uint256,
+            uint256,
+            uint256,
+            uint256,
+            address
+        )
+    {
         return (
             minimumContribution,
             address(this).balance,
@@ -94,7 +106,7 @@ contract Campaign {
         );
     }
 
-    function getRequestsCount() public view returns (uint) {
+    function getRequestsCount() public view returns (uint256) {
         return requests.length;
     }
